tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/bufferedAmount/bufferedAmount-large.html

<!doctype html>
<meta charset=utf-8>
<title>WebSockets: bufferedAmount for 65K data</title>
<meta name=timeout content=long>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  var data = "";
  ws.onopen = t.step_func(function(e) {
    for (var i = 0; i < 65000; i++) {
      data = data + "x";
    }
    ws.send(data);
    assert_equals(data.length, ws.bufferedAmount);
  });
  ws.onmessage = t.step_func(function(e) {
    assert_equals(e.data, data);
    t.done();
  })
});
</script>
